HUD ausgelagert in eigene Klasse
